demos/06_multimodal_voice/requirements.txt
langchain>=1.2.0
langchain-openai>=1.1.6
langchain-community>=0.4.1
python-dotenv==1.2.1
gradio>=5.0.0
edge-tts>=7.2.7
openai-whisper>=20250625
